+++ pkg/cmd/web.go
@@ -40,8 +40,12 @@ func newMacaron() *macaron.Macaron {
 		Delims:     macaron.Delims{Left: "[[", Right: "]]"},
 	}))
 
+	if setting.EnforceDomain {
+		m.Use(middleware.ValidateHostHeader(setting.Domain))
+	}
+
 	m.Use(middleware.GetContextHandler())
-	m.Use(middleware.Sessioner(setting.SessionOptions))
+	m.Use(middleware.Sessioner(&setting.SessionOptions))
 
 	return m
 }