+++ pkg/services/sqlstore/dashboard_test.go
@@ -51,6 +51,39 @@ func TestDashboardDataAccess(t *testing.T) {
 				So(query.Result.Slug, ShouldEqual, "test-dash-23")
 			})
 
+			Convey("Should return error if no dashboard is updated", func() {
+				cmd := m.SaveDashboardCommand{
+					OrgId:     1,
+					Overwrite: true,
+					Dashboard: map[string]interface{}{
+						"id":    float64(123412321),
+						"title": "Expect error",
+						"tags":  []interface{}{},
+					},
+				}
+
+				err := SaveDashboard(&cmd)
+				So(err, ShouldNotBeNil)
+			})
+
+			Convey("Should not be able to overwrite dashboard in another org", func() {
+				query := m.GetDashboardQuery{Slug: "test-dash-23", OrgId: 1}
+				GetDashboard(&query)
+
+				cmd := m.SaveDashboardCommand{
+					OrgId:     2,
+					Overwrite: true,
+					Dashboard: map[string]interface{}{
+						"id":    float64(query.Result.Id),
+						"title": "Expect error",
+						"tags":  []interface{}{},
+					},
+				}
+
+				err := SaveDashboard(&cmd)
+				So(err, ShouldNotBeNil)
+			})
+
 			Convey("Should be able to search for dashboard", func() {
 				query := m.SearchDashboardsQuery{
 					Title: "test",