+++ pkg/setting/setting.go
@@ -65,12 +65,14 @@ var (
 	RouterLogging      bool
 	StaticRootPath     string
 	EnableGzip         bool
+	EnforceDomain      bool
 
 	// Security settings.
 	SecretKey          string
 	LogInRememberDays  int
 	CookieUserName     string
 	CookieRememberName string
+	DisableGravatar    bool
 
 	// User settings
 	AllowUserSignUp    bool
@@ -85,6 +87,12 @@ var (
 	AnonymousEnabled bool
 	AnonymousOrgName string
 	AnonymousOrgRole string
+
+	// Auth proxy settings
+	AuthProxyEnabled        bool
+	AuthProxyHeaderName     string
+	AuthProxyHeaderProperty string
+	AuthProxyAutoSignUp     bool
 
 	// Session settings.
 	SessionOptions session.Options
@@ -348,16 +356,18 @@ func NewConfigContext(args *CommandLineArgs) {
 	Domain = server.Key("domain").MustString("localhost")
 	HttpAddr = server.Key("http_addr").MustString("0.0.0.0")
 	HttpPort = server.Key("http_port").MustString("3000")
-
 	StaticRootPath = makeAbsolute(server.Key("static_root_path").String(), HomePath)
 	RouterLogging = server.Key("router_logging").MustBool(false)
 	EnableGzip = server.Key("enable_gzip").MustBool(false)
+	EnforceDomain = server.Key("enforce_domain").MustBool(false)
 
 	security := Cfg.Section("security")
 	SecretKey = security.Key("secret_key").String()
 	LogInRememberDays = security.Key("login_remember_days").MustInt()
 	CookieUserName = security.Key("cookie_username").String()
 	CookieRememberName = security.Key("cookie_remember_name").String()
+	DisableGravatar = security.Key("disable_gravatar").MustBool(true)
+
 	// admin
 	AdminUser = security.Key("admin_user").String()
 	AdminPassword = security.Key("admin_password").String()
@@ -372,6 +382,13 @@ func NewConfigContext(args *CommandLineArgs) {
 	AnonymousEnabled = Cfg.Section("auth.anonymous").Key("enabled").MustBool(false)
 	AnonymousOrgName = Cfg.Section("auth.anonymous").Key("org_name").String()
 	AnonymousOrgRole = Cfg.Section("auth.anonymous").Key("org_role").String()
+
+	// auth proxy
+	authProxy := Cfg.Section("auth.proxy")
+	AuthProxyEnabled = authProxy.Key("enabled").MustBool(false)
+	AuthProxyHeaderName = authProxy.Key("header_name").String()
+	AuthProxyHeaderProperty = authProxy.Key("header_property").String()
+	AuthProxyAutoSignUp = authProxy.Key("auto_sign_up").MustBool(true)
 
 	// PhantomJS rendering
 	ImagesDir = filepath.Join(DataPath, "png")