+++ public/test/test-main.js
@@ -143,8 +143,9 @@ require([
     'specs/soloPanelCtrl-specs',
     'specs/table-specs',
     'specs/table-ctrl-specs',
+    'specs/singlestat-specs',
     'specs/dynamicDashboardSrv-specs',
-    'specs/unsavedChangesSrv-specs'
+    'specs/unsavedChangesSrv-specs',
   ];
 
   var pluginSpecs = (config.plugins.specs || []).map(function (spec) {